refactor(api): separa servicos por dominio e cria api client
