+++ front-end/src/features/companies/services/companies.api.ts
@@ -1,0 +1,4 @@
+import { createCrudApi } from '../../../shared/lib/api-client';
+import { Company } from '../types/company.types';
+
+export const companiesApi = createCrudApi<Company>('/api/companies');
