+++ front-end/src/features/contracts/services/contracts.api.ts
@@ -1,0 +1,4 @@
+import { createCrudApi } from '../../../shared/lib/api-client';
+import { Contract } from '../types/contract.types';
+
+export const contractsApi = createCrudApi<Contract>('/api/contracts');
