+++ front-end/src/features/expenses/services/expenses.api.ts
@@ -1,0 +1,4 @@
+import { createCrudApi } from '../../../shared/lib/api-client';
+import { Expense } from '../types/expense.types';
+
+export const expensesApi = createCrudApi<Expense>('/api/expenses');
