+++ front-end/src/features/freights/services/freights.api.ts
@@ -1,0 +1,4 @@
+import { createCrudApi } from '../../../shared/lib/api-client';
+import { Freight } from '../types/freight.types';
+
+export const freightsApi = createCrudApi<Freight>('/api/freights');
