+++ front-end/src/features/providers/services/providers.api.ts
@@ -1,0 +1,4 @@
+import { createCrudApi } from '../../../shared/lib/api-client';
+import { Provider } from '../types/provider.types';
+
+export const providersApi = createCrudApi<Provider>('/api/providers');
