+++ front-end/src/features/vehicles/services/vehicles.api.ts
@@ -1,0 +1,4 @@
+import { createCrudApi } from '../../../shared/lib/api-client';
+import { Vehicle } from '../types/vehicle.types';
+
+export const vehiclesApi = createCrudApi<Vehicle>('/api/vehicles');
